pyproject.toml
[project]
name = "docling-cvat-tools"
version = "0.1.0" # DO NOT EDIT, updated automatically
description = "CVAT annotation tools for Docling document processing and evaluation"
license = "MIT"
keywords = ["docling", "cvat", "annotation", "evaluation"]
readme = "README.md"
authors = [
    { name = "Peter Staar", email = "taa@zurich.ibm.com" },
    { name = "Panos Vagenas", email = "pva@zurich.ibm.com" },
    { name = "Ahmed Nassar", email = "ahn@zurich.ibm.com" },
    { name = "Nikos Livathinos", email = "nli@zurich.ibm.com" },
    { name = "Maxim Lysak", email = "mly@zurich.ibm.com" },
    { name = "Michele Dolfi", email = "dol@zurich.ibm.com" },
    { name = "Christoph Auer", email = "cau@zurich.ibm.com" },
]
classifiers = [
    "Operating System :: MacOS :: MacOS X",
    "Operating System :: POSIX :: Linux",
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Programming Language :: Python :: 3",
]
requires-python = '>=3.10,<4.0'
dependencies = [
    "docling-core>=2.48.0,<3.0.0",
    "docling>=2.56.1,<3.0.0",
    "pydantic>=2.0.0,<3.0.0",
    "lxml>=5.3.0,<6.0.0",
    "pandas>=2.2.3,<3.0.0",
    "openpyxl>=3.1.5,<4.0.0",
    "pillow>=10.3.0,<12.0.0",
    "tqdm>=4.67.1,<5.0.0",
    "typer>=0.16.0",
    "cvat-sdk>=2.40.0",
    'ocrmac>=1.0.0; sys_platform == "darwin" and platform_machine == "arm64"',
    "xlsxwriter>=3.1.0",
]

[project.urls]
homepage = "https://github.com/docling-project/docling-cvat-tools"
repository = "https://github.com/docling-project/docling-cvat-tools"
issues = "https://github.com/docling-project/docling-cvat-tools/issues"

[project.scripts]
docling-cvat-validator = "docling_cvat_tools.cli.cvat_validator_cli:main"
docling-cvat-to-docling = "docling_cvat_tools.cli.cvat_to_docling_cli:main"

[tool.uv]
package = true

[tool.setuptools.packages.find]
include = ["docling_cvat_tools*"]

[dependency-groups]
dev = [
    "pre-commit~=3.7",
    "mypy~=1.10",
    "black~=24.4",
    "isort~=5.10",
    "autoflake~=2.0",
    "flake8~=7.1",
    "flake8-docstrings~=1.6",
    "types-setuptools~=70.3",
    "pandas-stubs~=2.1",
    "types-openpyxl~=3.1",
    "types-requests~=2.31",
    "types-protobuf~=5.29",
    "boto3-stubs~=1.37",
    "types-urllib3~=1.26",
    "coverage~=7.6",
    "pytest~=8.3",
    "pytest-cov>=6.1.1",
    "pytest-dependency~=0.6",
    "pytest-xdist~=3.3",
    "python-semantic-release~=7.32",
    "python-dotenv~=1.0",
]

[tool.black]
line-length = 88
target-version = ["py310"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88
py_version = 310

[tool.mypy]
pretty = true
# strict = true
no_implicit_optional = true
plugins = "pydantic.mypy"
python_version = "3.10"
exclude = '^\\.venv/'

[[tool.mypy.overrides]]
module = [
    "docling_core.*",
    "docling.*",
    "cvat_sdk.*",
    "lxml.*",
    "PIL.*",
    "xlsxwriter.*",
    "ocrmac.*",
]
ignore_missing_imports = true
follow_imports = "skip"

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "E501"]

[tool.semantic_release]
# for default values check:
# https://github.com/python-semantic-release/python-semantic-release/blob/v7.32.2/semantic_release/defaults.cfg

version_source = "tag_only"
branch = "main"

# configure types which should trigger minor and patch version bumps respectively
# (note that they must be a subset of the configured allowed types):
parser_angular_allowed_types = "build,chore,ci,docs,feat,fix,perf,style,refactor,test"
parser_angular_minor_types = "feat"
parser_angular_patch_types = "fix,perf"

